% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLambdaGrid.R
\name{createLambdaGrid}
\alias{createLambdaGrid}
\title{Creates Lambda Solution Path for crrBAR}
\usage{
createLambdaGrid(ftime, fstatus, XX, uuu, lam.min, nlambda, log = TRUE)
}
\arguments{
\item{ftime}{A vector of event/censoring times.}

\item{fstatus}{A vector with unique code for each event type and a separate code for censored observations.}

\item{XX}{Standardized verion of \code{X} (already standardized in \code{crrBAR})}

\item{uuu}{Vector of IPCW weights calculated in \code{crrBAR}}

\item{lam.min}{Numeric: smallest value of lambda if performing grid search}

\item{nlambda}{Numeric: number of \code{lambda} values if performing grid search  (default is 25)}

\item{log}{Logical: Whether or not the grid search is log10 spaced (default is \code{TRUE})}
}
\description{
Creates a grid of \code{lambda} for \code{crrBAR}. This is an internal function.
}
\details{
Not intended for use by users. This function is called in \code{crrBAR} if \code{lambda} is missing.
}
