% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLogLikelihood.R
\name{getLogLikelihood}
\alias{getLogLikelihood}
\title{Get Log-Pseudo Likeihood}
\usage{
getLogLikelihood(ftime, fstatus, X, failcode = 1, cencode = 0,
  beta = rep(0, dim(X)[2]))
}
\arguments{
\item{ftime}{A vector of event/censoring times.}

\item{fstatus}{A vector with unique code for each event type and a separate code for censored observations.}

\item{X}{A matrix of fixed covariates (nobs x ncovs)}

\item{failcode}{Integer: code of \code{fstatus} that event type of interest (default is 1)}

\item{cencode}{Integer: code of \code{fstatus} that denotes censored observations (default is 0)}

\item{beta}{A vector of coefficient values at which to evaluate log-pseudo likelihood}
}
\value{
Returns log-pseudo likelihood at \code{beta}.
}
\description{
Returns log-pseudo likeihood at a specific value of \code{beta}.
}
\examples{
library(cmprsk)
set.seed(10)
ftime <- rexp(200)
fstatus <- sample(0:2, 200, replace = TRUE)
cov <- matrix(runif(1000), nrow = 200)
dimnames(cov)[[2]] <- c('x1','x2','x3','x4','x5')
fit <- crr(ftime, fstatus, cov)
fit$loglik
getLogLikelihood(ftime, fstatus, cov, beta = fit$coef)
}
\references{
Fine J. and Gray R. (1999) A proportional hazards model for the subdistribution of a competing risk.  \emph{JASA} 94:496-509.
}
